Initialize core_gossip_message_validator module by extracting CoreGossipMessageValidator from core_executor
