+++ domains/client/domain-executor/src/lib.rs
@@ -66,6 +66,7 @@ mod core_bundle_processor;
 mod core_bundle_producer;
 mod core_domain_worker;
 mod core_executor;
+mod core_gossip_message_validator;
 mod domain_block_processor;
 mod domain_bundle_producer;
 mod domain_bundle_proposer;
@@ -81,9 +82,8 @@ mod system_gossip_message_validator;
 mod tests;
 mod utils;
 
-pub use self::core_executor::{
-    Executor as CoreExecutor, GossipMessageValidator as CoreGossipMessageValidator,
-};
+pub use self::core_executor::Executor as CoreExecutor;
+pub use self::core_gossip_message_validator::CoreGossipMessageValidator;
 pub use self::system_executor::Executor as SystemExecutor;
 pub use self::system_gossip_message_validator::SystemGossipMessageValidator;
 use sc_utils::mpsc::TracingUnboundedSender;
